requirements.txt
streamlit==1.45.1
numpy==2.2.6
pillow==11.2.1
plotly==5.19.0
streamlit-option-menu==0.4.0 